separation of all of the algoritms in files and folders; added softmax activation function; added iris flower dataset example
