+++ .gitignore
@@ -12,3 +12,7 @@ logs
 results
 
 npm-debug.log
+
+node_modules
+*.iml
+*.idea
