+++ README.md
@@ -39,15 +39,15 @@ var XOR_IDEAL = [
     [0]
 ];
 
-var network = ENCOG.BasicNetwork.create( [
-    ENCOG.BasicLayer.create(ENCOG.ActivationSigmoid.create(), 2, 1),
-    ENCOG.BasicLayer.create(ENCOG.ActivationSigmoid.create(), 3, 1),
-    ENCOG.BasicLayer.create(ENCOG.ActivationSigmoid.create(), 1, 0)
-    ] );
+var network = ENCOG.networks.basic.create([
+    ENCOG.layers.basic.create(ENCOG.activationFunctions.sigmoid.create(), 2, 1),
+    ENCOG.layers.basic.create(ENCOG.activationFunctions.sigmoid.create(), 3, 1),
+    ENCOG.layers.basic.create(ENCOG.activationFunctions.sigmoid.create(), 1, 0)
+]);
 
 network.randomize();
 
-var train = ENCOG.PropagationTrainer.create(network, XOR_INPUT, XOR_IDEAL, "RPROP", 0, 0);
+var train = ENCOG.trainers.propagation.create(network, ENCOG.errorFunctions.linear.create(), XOR_INPUT, XOR_IDEAL, "RPROP", 0, 0);
 
 var iteration = 1;
 
@@ -58,17 +58,17 @@ do {
     iteration++;
 } while (iteration < 1000 && train.error > 0.01);
 
-var input = [0,0];
+var input = [0, 0];
 var output = [];
 
 console.log("Testing neural network: \n");
 
-for (var i=0;i<XOR_INPUT.length;i++) {
-    network.compute(XOR_INPUT[i],output);
+for (var i = 0; i < XOR_INPUT.length; i++) {
+    output = network.compute(XOR_INPUT[i]);
     var testResultString = "Input: " + String(XOR_INPUT[i][0]) +
-          " ; " + String(XOR_INPUT[i][1]) +
-          "   Output: " + String(output[0]) +
-          "   Ideal: " + String(XOR_IDEAL[i][0]);
+        " ; " + String(XOR_INPUT[i][1]) +
+        "   Output: " + String(output[0]) +
+        "   Ideal: " + String(XOR_IDEAL[i][0]);
     console.log(testResultString + "\n");
 }
 ```
@@ -98,7 +98,7 @@ The included XOR example can be simply run by,
 ```js
 var ENCOG = require('encog-node');
 
-ENCOG.example();
+ENCOG.examples.xor();
 ```
 
 This will run the same XOR example mentioned above.
