+++ examples/xor.js
@@ -3,9 +3,9 @@
  * @author Jeff Heaton
  * @source http://www.codeproject.com/Articles/477689/JavaScript-Machine-Learning-and-Neural-Networks-wi
  */
-var ENCOG = require('../lib/encog-js');
+var ENCOG = require('../index');
 
-module.exports = function() {
+module.exports = function () {
     var XOR_INPUT = [
         [0, 0],
         [1, 0],
@@ -20,15 +20,15 @@ module.exports = function() {
         [0]
     ];
 
-    var network = ENCOG.BasicNetwork.create( [
-        ENCOG.BasicLayer.create(ENCOG.ActivationSigmoid.create(), 2, 1),
-        ENCOG.BasicLayer.create(ENCOG.ActivationSigmoid.create(), 3, 1),
-        ENCOG.BasicLayer.create(ENCOG.ActivationSigmoid.create(), 1, 0)
-        ] );
+    var network = ENCOG.networks.basic.create([
+        ENCOG.layers.basic.create(ENCOG.activationFunctions.sigmoid.create(), 2, 1),
+        ENCOG.layers.basic.create(ENCOG.activationFunctions.sigmoid.create(), 3, 1),
+        ENCOG.layers.basic.create(ENCOG.activationFunctions.sigmoid.create(), 1, 0)
+    ]);
 
     network.randomize();
 
-    var train = ENCOG.PropagationTrainer.create(network, XOR_INPUT, XOR_IDEAL, "RPROP", 0, 0);
+    var train = ENCOG.trainers.propagation.create(network, ENCOG.errorFunctions.linear.create(), XOR_INPUT, XOR_IDEAL, "RPROP", 0, 0);
 
     var iteration = 1;
 
@@ -39,17 +39,17 @@ module.exports = function() {
         iteration++;
     } while (iteration < 1000 && train.error > 0.01);
 
-    var input = [0,0];
+    var input = [0, 0];
     var output = [];
 
     console.log("Testing neural network: \n");
 
-    for (var i=0;i<XOR_INPUT.length;i++) {
-        network.compute(XOR_INPUT[i],output);
+    for (var i = 0; i < XOR_INPUT.length; i++) {
+        output = network.compute(XOR_INPUT[i]);
         var testResultString = "Input: " + String(XOR_INPUT[i][0]) +
-              " ; " + String(XOR_INPUT[i][1]) +
-              "   Output: " + String(output[0]) +
-              "   Ideal: " + String(XOR_IDEAL[i][0]);
+            " ; " + String(XOR_INPUT[i][1]) +
+            "   Output: " + String(output[0]) +
+            "   Ideal: " + String(XOR_IDEAL[i][0]);
         console.log(testResultString + "\n");
     }
 };
